+++ packages/larva-patterns/components/c-input/c-input.prototype.js
@@ -4,6 +4,5 @@ module.exports = {
 	c_input_label_class: 'lrv-a-screen-reader-only',
 	c_input_label_text: 'Input Example',
 	c_input_class: '',
-	c_input_name: 'input_example',
 	c_input_placeholder_attr: 'Enter text.',
 };